Naved Asif/AIQUIZCRAFT/requirements.txt
pandas
numpy
python-dotenv
PyPDF2
langchain
openai
flask
requests


